requirements.txt
streamlit
google-genai
python-dotenv
pandas
openpyxl